Update cpp-marketplace-tab.md
The category and industry taxonomy structure for Azure Marketplace and AppSource is changing on 01/31/20. I am updating the marketplace publisher guide to make sure all sections are updated to the new taxonomy structure accordingly. Starting on 01/31/20, publishers need to select "IoT Edge Modules" as the only subcategory for this offer type.