Updating for new Bring your own storage feature
Azure Managed Applications now support "Bring your own storage" for use with Azure Service Catalog when creating managed applications definitions. I have updated the documentation to contain additional steps for utilizing this, as well as a note that it can be used in conjunction with Azure KeyVault to allow for customer managed keys.