add important requirement
Added Important note to requirements section to make the customer aware that they must have an Azure AD Joined Machine to RDP to a windows VM with Azure AD Auth enabled.

The missing requirement caused myself and several other Azure engineers to fail to get this working several times. 
There is a similar not much farther down in the document, but putting all requirements in the requirements section makes it easier for customers to get this working.