custom parameters definition location in CI/CD
During git integration setup, user chooses a path/folder to be treated as root.  This is not necessarily the same as the root of the repository.  The location of custom parameters definition should reflect this nuance.
Do you think this became too wordy?
-------
cc: @djpmsft
GitHub Issue: https://github.com/MicrosoftDocs/azure-docs/issues/46149