Subscriptions that went through directory changes
For subscriptions that went through directory changes, customers will by default land in the new directory after signing into the Azure portal. From their perspective, they should manage everything of the subscriptions in the new directory. This is true for the Subscriptions blade. However, they may find that they are unable to find these subscriptions when using the Cost Management + Billing blade, because this blade retrieves information by making calls to the Commerce Platform, and these subscriptions are still under the old directory in Commerce. This behavior is as expected. Customers will need to switch directory to the old directory before the subscriptions are shown correctly on the Cost Management + Billing blade. We should show a warning message here.