I added this comment in Additional considerations
I added this comment in Additional considerations

### Storage
Customer controlled failover (preview) only fails over the data plane. Control plane is not failed over, so the account continues to be managed by the instance of SRP located in the former primary region. This is why “location” does not change (it indicates the control plane location of the account, which ARM uses to route requests to the appropriate regional instance of SRP). Only “primaryLocation” and “secondaryLocation” are swapped, in order to indicate the physical location of the data. you can find this design when you deploy a HDInsight cluster using the failed-over storage.