Update how-to-manage-ua-identity-cli.md
Our documentation lacks clarity regarding the needs for providing permissions to Graph API for corresponding service principals when managing users via CLI. I added a note in the prereq section however, if there is a better place for this information we should place it there. This is following direct customer feedback.