Clarify what the SDK is for
I found this doc and was confused by it. I didn't see any of these operators on the Cosmos DB JavaScript SDK. Someone had to explain to me that this API is only available in Stored Procedures and Triggers.

I think we can clarify the title and intro paragraph a bit to make that clear. When I read "server-side", I thought, "Well, I'm using Node. That's server-side. Why can't I use this?".

-------
cc: @markjbrown