Adding the requirement for a CNAME DNS record
I was using A records for DNS for my free app service certificate. It wasn't abundantly clear from the documentation that a CNAME is required (although it is mentioned in the docs), so I added a line in the section that explains HOW to create it that A records aren't supported.