(AzureCXP) fixes MicrosoftDocs/azure-docs#46755
The initial Virtual WAN Tutorial has some typos that seem to have trickled down into the How-To Guides. The original diagram/figure shows "VNetB" 10.3.0.0/16 but in various explanatory text and example configs, that has become 10.30.0.0/16

For this page, the example is still valid (it just doesn't agree with the original Tutorial and diagram). There may be other Tutorial and How-To pages with this same mismatch.