Grammar, typos, and general formatting
I corrected some grammar as well as some typos and/or missing context words. Also made the formatting consistent for the tables. Added bold around the word "unprotected" in the first paragraph because that is key information for the reader to know about the status of a failed over VM.