(AzureCXP) fixes MicrosoftDocs/azure-docs#47236
Replaced the description of Tenant parameter under ### Request the permissions from a directory admin section from "The directory tenant that you want to request permission from. Can be provided in GUID or friendly name format OR generically referenced with `common` as seen in the example." to "The directory tenant that you want to request permission from. Can be provided in GUID or friendly name format OR generically referenced with organizations as seen in the example. Do not use 'common', as personal accounts cannot provide admin consent except in the context of a tenant. To ensure best compatibility with personal accounts that manage tenants, use the tenant ID when possible."

-------
cc: @rwike77