Make this document easier to understand
Having just done this, I struggled for 2 days trying to figure out how the parameters worked. The concept of referencing the Key Vault is very straight forward. It's how everything fits together that is difficult to understand. Especially since the majority of the other documentation calls for using a parameters file.

Originally it says you must use **linked** templates which is not true, you can also use nested. And I think understanding nested templates makes linked templates simple. Also, by providing a complete template instead of just a resource block it shows how the key vault parameters are not defined in the top parameters block but they are defined in the resources.deployment.parameters block.