(Azure CXP) fixes doc update.
Updated Line 21 from 

`A multitenant application is a shared resource that allows separate users, or "tenants," to view the application as though it was their own. A typical scenario that lends itself to a multitenant application is one in which all users of the application may wish to customize the user experience but otherwise have the same basic business requirements.` 

to 

`A multitenant application is a shared resource that allows "users in separate tenants" to view the application as though it was their own. A typical scenario that lends itself to a multitenant application is one in which all users of the application from different tenants may wish to customize the user experience but otherwise have the same basic business requirements.`

Fixes MicrosoftDocs/azure-docs#47541

@wadepickett could you please review this and approve the same accordingly ? 

-------
cc: @wadepickett
GitHub Issue: https://github.com/MicrosoftDocs/azure-docs/issues/47541