Improve Advice on Avoiding OS Version Changes
Excluding the release files doesn't actually stop the OS from upgrading-- it just makes it report an incorrect version. The kernel and all supporting programs that make a functionally different OS version are still installed. Only the /etc/redhat-release file and other informational files are skipped.

Here is the contents of that RPM:

```text
[root@rhel7 tmp]# rpm -q -l redhat-release-server
/etc/issue
/etc/issue.net
/etc/os-release
/etc/pki/product-default
/etc/pki/product-default/69.pem
/etc/pki/rpm-gpg
/etc/pki/rpm-gpg/RPM-GPG-KEY-redhat-beta
/etc/pki/rpm-gpg/RPM-GPG-KEY-redhat-legacy-former
/etc/pki/rpm-gpg/RPM-GPG-KEY-redhat-legacy-release
/etc/pki/rpm-gpg/RPM-GPG-KEY-redhat-legacy-rhx
/etc/pki/rpm-gpg/RPM-GPG-KEY-redhat-release
/etc/redhat-release
/etc/rpm/macros.dist
/etc/system-release
/etc/system-release-cpe
/usr/lib/systemd/system-preset/85-display-manager.preset
/usr/lib/systemd/system-preset/90-default.preset
/usr/share/doc/redhat-release/GPL
```
RedHat EUS is a better option for version-locking RedHat since this actually avoids upgrading to the new software.