Clarify a proper WSFC supported solution
As is, this does NOT create a supported solution. Cluster validation is required to be run and pass (even if it is with warnings) to have a supported solution. This article has to reflect that or be pulled until such time as that is supported. Thanks.

Here is the documentation:
https://docs.microsoft.com/en-us/windows-server/failover-clustering/create-failover-cluster#validate-the-configuration
"Microsoft supports a cluster solution only if the complete configuration passes all validation tests and if all hardware is certified for the version of Windows Server that the cluster nodes are running."

Allan