Update maxient-conduct-manager-software-tutorial.md
This tutorial needed to be edited to match Maxient's operating procedures for setting up Campus Based Authentication.

Maxient clients, set up Azure AD for use with Maxient.  They must allow ALL USERS to authenticate (this does not grant access to the Maxient software) we ONLY use Azure to *authenticate* we take care of the *authorization* on our end.   Once things are set up, the client sends us their metadata URL and we reply with a test link.  This test link is an attribute reflector.  We need to make sure the proper attributes are being released.  Once that is done - the setup is complete.  We will turn on Authentication within Maxient when appropriate (immediately for current clients - just before training for clients in the process of implementation).   I believe the edits I made to this document reflect this, but it will need a little clean up (I am certain i messed up some tags and what not).  All references to a "test user" must be removed...that is not how our process works.

Any questions, please let me know.   Thanks!