Update limit-egress-traffic.md
We're getting some cases with customers blocking port 10250 - and facing Helm or Metrics issue.
I can't find any public doc with the port 10250 needed to be opened - can we add it somewhere? this doc seems to be right place

"For communication between kubelet and metrics server, before AKS 1.16, 10255 port on the node is required to be open. After 1.16, port 10250 is required. Port 10250 is also required for other functionality of kubelet."