dbo.vTableSizes view not accurate for small SLO
The view definition for dbo.vTableSizes doubles the sizes of the tables when the DW size (SLO) is DW500c or below.  This is happening because the query assumes that the control node is separate from the compute nodes; however, on lower SLOs, it is not.  As such, a filter needs to be added to ensure the counts are accurate.