Including 'spring-cloud-azure-appconfiguration-config-web' dependency
It would be very beneficial for developers to include 'spring-cloud-azure-appconfiguration-config-web' in this doc as the out of the box feature of auto refresh would be handy for any application. These details are available under - https://github.com/microsoft/spring-cloud-azure/tree/master/spring-cloud-azure-starters/spring-cloud-starter-azure-appconfiguration-config#dependency-management . But it would also make more sense to include this dependency in this quick start java spring documentation.