Updating "Getting started" with additional step needed to install git lfs
Previously the last instruction on installing Git just said 'brew install git-lfs', but an additional step is needed to finish installation (running 'git lfs install' or 'git lfs install --system). This is a really easy step to miss (the instruction to do it is hidden in the output of 'brew install git-lfs'), and results in all sorts of unexpected errors popping up later (i.e. AppIcon images not working properly).