Update marketplace-publishers-guide.md
The deleted statement is causing confusion among publishers, because they don't know where to go because there is no instruction in this document regarding where they should go to "...send us a message under Problem Type > Marketplace Onboarding.  Our Marketplace Publisher Onboarding Team will be happy to help!"

We want to remove this statement now, with the idea that we will update this document with information on where to go to get publisher/partner support. 

Thanks!