Clarify PublicIPAddressLimitReached errors for Databricks
Microsoft Support has recently received multiple cases in which the customer was confused why they continued to experience PublicIPAddressLimitReached errors when attempting to create a new Databricks cluster after recently terminating other clusters. When Microsoft Support explains why this may be expected behavior and suggests to request an increase in the corresponding Azure resource quota, the customer frequently becomes irate that the expected behavior is not publicly documented and that there is not another solution. This commit is intended to prevent such issues by diplomatically documenting that behavior and also suggesting multiple alternative solutions.