Behavior of unhealthy threshold with v2 is different from v1.
I found that when health probe was failed in AppGW v2, next probe was sent after 30s interval. I think the following sentence in this document is only v1 SKU.  Could you revi

"These additional health probes are sent in quick succession to determine the health of the backend quickly and don't wait for the probe interval. "