Fix Redis usage sample code
This sample code was using the wrong variable to refer to the ConnectionMultiplexer object that is used for communicating with Redis. It is clear from the comments in the sample as well as the code snippet above this sample that the original intention was to use the "Connection" property instead of the private "lazyConnection" field.