Clarification of NotActions and multiple roles
I'd like to see the not cleaned up in the NotActions section to be a little more clear about what happens when multiple roles are assigned.

The way that I read this is if one roles has NotActions but a second role has Actions that the Actions allowed on the second role would override the NotActions specified on the first role. This appears to be the case only if the Actions are specified in the second role, not wildcarded.

For an example, take Contributor and Owner. They both have <"actions": ["*"],> but the Contributor role has specific NotActions defined. Those specified NotActions appear to override the wildcard allowed actions on the Owner role.

I've simply changed language here to say that a second role that grants explicit access to the same operations (not wildcarded) will allow the user to perform that operation.