Eliminsted the duplicate VM Size section
As reported in Github issue 45597, the VM Size section was duplicated in two places with the exact same text.  I just deleted the second one since it was an exact duplicate.

Also on the first on, I fixed the casing on the section title to "VM Size" from "Vm size" so it matches everything else.