Edit variables.tf
Got this error when applying Terraform config:  "Terraform 0.11 and earlier required type constraints to be given in quotes,
but that form is now deprecated and will be removed in a future version of
Terraform. To silence this warning, remove the quotes around "map" and write
map(string) instead to explicitly indicate that the map elements are strings."

Fixed by removing quotes around map string.