Incorrect reference to PS script at line 195, 198
The name of the blob context is $ctx whereas it is referenced as $context at line number 198. The name of the variable should be $ctx at line 198.

Likewise, the name of the PS variable for blob is $blobs at lin number 198, whereas it is referenced as $blob at line 198 should be $blob.