Update required in line number 23
In this section, we say that the blob if deleted cannot be recovered. This is only true data lake scenarios where there is no concept of delete markers or versioning. For Azure blob store with soft delete enabled, the deleted blob can be undeleted and recovered.

We need to change this section and make this more explicit for the reader.