(AzureCXP)MicrosoftDocs/azure-docs/issues/51420
(AzureCXP) fixed 9th step of Expose an API > Set  in configuring with advanced settings.  
Done repro with previous instructions and found error with multi-tenant app although fine with the single-tenant.
below is the doc enhancement done :-
For single-tenant app, paste in the URL of your App Service app and select Save.
For multi-tenant app, paste in the URL which is based on one of tenant verified domains and select Save.