(Azure CXP) Technical Doc Update
Added the following at line 37 . 

2. [License Requirements](#license-requirements)

and the following at line 47

## License Requirements

In order to access the sign-in reports for a tenant, an Azure AD tenant must have associated Azure AD Premium license. Azure AD Premium P1 (or above) license is required  to access sign-in reports for any Azure AD tenant. Alternatively if the directory type is Azure AD B2C , the sign-in reports are accessible through the API without any additional license requirement. 


As per MicrosoftDocs/azure-docs#51693

-------
cc: @MarkusVi