(Azure CXP) Updated as per
Updated Line 134 because the output log file location would be seen in present working directory where we run the script rather than  `%WinDir%/System32/IoTAgentLog.log`

from 

`1. Review the log file for more information about the failure.

   Log file location: `%WinDir%/System32/IoTAgentLog.log` 
`
to 

`1. Review the log file for more information about the failure. The log file would be present in the working directory where we run the script. 

   Log file location: `.\IoTAgentLog.log`
`
This is as per MicrosoftDocs/azure-docs#51943
-------
cc: @mlottner