Update doc apache-hive-migrate-workloads to improve internal metastore export script and clarify what steps are required for internal vs external metastore migration.
Certain steps are necessary only after restore of external metastore. Restructuring of sections in this update clarifies this.
This update also makes minor sentence changes for clarity.
bash command to generate alltables.hql moves to a script file in Storage Account. We may need a way to store backups.
Several updates to this alltables generation script, including the following cases:
  - adjusts easily for ESP
  - column names like `date`, which require backticks
  - no more name conflicts for temporary staging tables
  - fileformat compatible for deep copy of ACID tables
  - copies empty partitions