Update SQL VM registration documentation to incorporate Full mode installation without Lightweight mode installation
It is feasibile to register the SQLVM Provider directly to Full mode without undertaking the Lightweight mode installation. Of course, a SQL Server service, or the VM, restart is required, and should be planned according to company scheduled maintenance windows. This update to the documentation is to provide the PowerShell and Bash commands to directly perform a Full mode installation.