[Do not merge until Build] Update application settings reference
Very briefly adds the new setting and fixes the default values (which are false not true).

I don't add a section laying out which parts of the word is removed because its relative to the word, not every word will have 's' removed from it, i.e. 'this' won't be 'thi', but 'pictures' will be 'picture' provided it exists as a word underneath the hood in LUIS which we don't really expose, I'm happy to keep this vague for now as it is a soft rule not a hard one. Let me know what you think.