Some minor changes
Added mention of Delta Lake support.
Updated netcoreapp3.0 to netcoreapp3.1 (please check that it works :))
moved the content for On Linux section to correct place.
Removed duplicate content
Is the documentation link for the .NET for Apache Spark documentation working?

It would be great to add some examples for each of the features that are called out. Can we add that in future doc refreshes please?