Update event-hubs-limits.md
the first set of limits do not apply to Dedicated Tier, for example the limit of 10 Event Hubs per Namespace. The table below says for Dedicated you can have 1000 EHs per namespace. Also the other values in that first table to not seem to apply to dedicated (please correct me if I'm wrong).