[Event Hubs] Send/Receive Quickstart Tweaks
The purpose of these changes is to adjust some verbiage to better reflect
the prerequisites needed for the quickstart, clarifying that the SDK tooling is
being discussed and not the target framework of the output.

Additionally, updated mentions of `.NET Core` to `.NET` to reflect that the
library may be used to build both .NET Core and .NET framework applications and
is the recommended Event Hubs library for both.