Add Azure REST API reference
The document mainly refers to the change of DNS aliases with PowerShell, leaving aside the references to the Azure REST API that also allows to perform the same actions. The two links included allow you to learn about the authentication process with the Azure REST API in general and the methods of said API that allow you to modify DNS aliases.