The Circuit Breaker resource is titled incorrectly
The circuit breaker takes advantage of a special flavor of stateful implementation based on the entity pattern. I am of the opinion that rather generally calling the same Durable *Functions*, it should explicitly point out that is an implementation of the Durable **entity** concept, which can even be used without using a Durable Functions orchestrator (which, otherwise, the current title implies).