Update get-started-model-ingestion.md
The terms 3D CAD Drawing, 3d model, 3d asset and 3d file seem to be used interchangeably. All these terms are similar, but could refer to different things. In this PR I'm proposing we use only the term `3D model` throughout the documentation.

I also removed the AssetId field because it is not used anymore