Preventing DNS resolution trough AzurePlatformDNS
Article says "By default DNS communication is not subject to the configured network security groups unless specifically targeted leveraging the [AzurePlatformDNS](../virtual-network/service-tags-overview.md#available-service-tags) service tag."

When trying to implement this, it only allows to specify "*" as destination port. Trying to use port 43 or TCP/UDP protocols causes a generic error setting up rule ("An error ocurred"). It also only accepts "*" as destination port and "Any" as protocol, this time with a more specific error: "DNS exfiltration tags must be used with "*" as a protocol."