Updated to address possible customer confusion
https://github.com/Azure-Samples/cognitive-services-speech-sdk/issues/682
There was potential for customer confusion here, as we suggested customers can merge multiple LUIS intent responses. But we had no easy way to do that or to give customers instructions for this. So I removed that detail and set clear expectation for what customers would get for longer text submissions to LUIS>