Include Additional Scope Options
I spent a good amount of time troubleshoot this for a Go service I'm writing to hook into Azure AD. Finally was able to trace out requests in the MSAL app and see that the Python version actually just sends a non-encoded scope list to the authorize and token endpoints. This caused a good deal of grief so I thought it would be prudent to at least mention it in the docs. Not sure how great the copy is, so definitely would like some input on how I worded it. 

Thanks y'all and keep up the great work!