(AzureCXP) Updating TimeStampField in PS Script
Based on customer feedback on below GitHub thread , validated the PS script . The script what we is only creating the table in Log Analytics workspace but not ingesting data. However the C# code works , after closing looking into it PowerShell script $TimeStampField = "DateValue" is causing the issue. Not sure what API changes have occurred which affected this but changing it to  $TimeStampField = "" is making it work. 

https://github.com/MicrosoftDocs/azure-docs/issues/57052