Final tweaks
Removed autoprovisioning because it's not available until July

From AKS FAQ:
Autoprovisioning is available for the Log Analytics agent to all your nodes, including those that are hosted on virtual machine scale sets.

From Threat Protection include:
The deployment of the agent can be automated with Security Center's autoprovisioning option.