Update tutorial-ingress-controller-add-on-existing.md
Previously I updated the document a bit hastily but looking at it further, it apparently seems like creating a registration with name "microsoft.containerservice" irrespective of making it "Microsoft.ContainerService".  No clue why this is.  But then the document should reflect the lower case of the namespace "microsoft.containerservice" while querying for the status for the registration or else you don't get any status result.  

Can this be made better?  Why is the name space "microsoft.containerservice" in lowercase while all others are in pascal casing?