Update VM Details - Clarification around OU format
On step 10, I've added a new note to inform readers that when typing the Organizational Unit for domain join, the OU must be written in DN format and must also not include quotations. If it does include quotations (which occurs if copied from Active Directory Users + Computers or Active Directory Administrative Centre) the join domain phase will fail, resulting in a session host which is not registered to the app group as it previously failed during domain join. This saves the potential pain of having to then delete the VM and its associated resources and re-deploy.