Updates Review and Create Section to fix yml syntax
I copied the instructions in this article and ran into an issue with the build and deploy of the site to my static webapp as the Github action was complaining about .yml syntax errors on those lines. changing the indentation to the proposed changes fixed the error and my site successfully built.