About force tunneling for V1.
I read below sentence but it was too difficult to understand whether  force tunneling is supported by AppGW v1 or not. I hope you fix this document.
 
"Failure to do so might result in incorrect health-probe or traffic-routing behavior. This includes learned routes or default 0.0.0.0/0 routes that are propagated by Azure ExpressRoute or VPN gateways in the virtual network."