Fix typo in links to EventHub RBAC role definition
This pull request fixes a typo in the documentation for the EventHub RBAC role definitions, where the link for the "Azure Event Hubs Data sender" entry links to the receiver and the link for the "Azure Event Hubs Data receiver" entry links to the sender. This pull request makes the links consistent such the sender entry links to the sender and the receiver entry links to the receiver.