Update groups-dynamic-membership.md
Removed the false licensing requirement.  You can say it's like that all you want, but it isn't.  There's no point in documenting something that isn't true.  If and when you implement these controls into the Dynamic Group functionality, then it will be fine to document it.  Until then, this note has no validity.