Changed sentence
In the following paragraph, it should be "returned" instead of "returned back".

On the sign-up or sign-in page, Azure AD B2C presents a list of external identity providers the user can choose for sign-in. Once they select one of the external identity providers, they're taken (redirected) to the selected provider's website to complete the sign in process. After the user successfully signs in, they're returned back to Azure AD B2C for authentication of the account in your application.
-------
cc: @msmimart
GitHub Issue: https://github.com/MicrosoftDocs/azure-docs/issues/59682