Add viewport meta tag into the sample HTML
The proposed change here will allow the page to zoom on mobile devices and allow CSS media queries to work.

Its good practice to add this tag into HTML5 pages to support accessible zooming and media queries in CSS.

I took the example on this page as a basis to try and create a custom user flow page, and then used it to add more and more styling. Took me quite a while to figure out why CSS media queries weren't working as I just assumed this was there.