Updated to username/password issue
Added an important section reminding customers that Azure AD Connect needs to be joined to the domain.  Reason for adding, worked an SR with a customer where customer and I saw article and still were not able to figure out issue.  This will prevent SRs for customer's who do not read the prerequisites for Azure AD Connect.