Fixed language about --generate-ssh-keys behavior when keys already exist
This change is to address GitHub issue 45043

Previously the article indicated an error was thrown if the key pair already existed.  That may have been true at one time, but today at least, the command will just use the existing keys (which as the issue author points out is the desired behavior).  Updated language to account for this.  Also updated command to include --image option because that is required to actually create the VM.  Now a user can copy and paste