(Azure CXP) UPdate Thumbnailphoto information.
Added description on Line 163 , 245 , 296 as per MicrosoftDocs/azure-docs#59792

"synced only once from Azure AD to Exchange Online after which Exchange Online becomes source of authority for this attribute and any later changes can't be synced from on-premise. See ([KB](https://support.microsoft.com/help/3062745/user-photos-aren-t-synced-from-the-on-premises-environment-to-exchange)) for more." as per article. https://support.microsoft.com/en-us/help/3062745/user-photos-aren-t-synced-from-the-on-premises-environment-to-exchange

Closes MicrosoftDocs/azure-docs#59792


-------
cc: @billmath