Update app-attach.md
fixed the stage.ps1 script. Some of the commands were splits across multiple lines which failed when you executed the copy pasted script.
Also updated the use of Mount-Vhd to Mound-Diskimage which is a widely adopted practice. This removes the need for the Hyper-V management tools to be installed on the session hosts.