(forumsteam) fixes MicrosoftDocs/azure-docs#60410
The doc said that if the files exceed the local cache value, then local cache will be ignored and use the file system instead. This is not entirely true as the only time local cache is ignored is if the user has defined the limit to 2 GB AND their deployment files exceed 2 GB.

Updated the doc to say, "If you've increased local cache limit to 2 GB and the copied files exceed the maximum size of 2 GB, App Service silently ignores local cache and reads from the remote file share. If there is no limit defined or the limit is set to anything lower than 2 GB and the copied files exceed 2 GB, the deployment or swap may fail with an error."