Update ClientId settings
Here is what I have found. Scenario: SPA using JWT token to authenticate an API. The spa gets the token successfully, inside we have all information required, the audience (aud) is: "api://clientid-guid". But then with the basic configuration, JWT validations fails and the error said "audience incorrect 'api://clientid-guid' ". Adding Audience parameter to appsettings.json did not help. The only thing that worked for me was to change in appsettings.json ClientId to: "api://clientid-guid". For me, it looks like it took ClientId as valid Audience. I'm not sure this behavior is expected or not. Cheers.