Clarify x5t header calculation specification
While trying this out myself it was not obvious to see that it is not the Base64 of the Thumbprint string, but rather of the Hexadecimal representation of it. After some trial and error I figured it out. 
This is the node.js based code for generating the correct value from the Thumbprint:
```js
var thumbprintBase64 = Buffer.from('84E05C1D98BCE3A5421D225B140B36E86A3D5534', 'hex').toString('base64');
```