Added RBAC scope
I tested the permissions at the resource scope, with the principle of least privilege in mind, and received an error stating the resource group does not exist.  Once I gave the user principal the contributor role, scoped at the resource group, I was able to successfully run the Join-AzStorageAccountForAuth cmdlet.