Update concepts-convention.md
Clarify it's not a requirement that the acknowledgement for a writeable property update reports the same value as the received value. Typically that is the case, but ultimately it's up to the device app to decide what is the right value to report. For example if the desired value is actually invalid then the device would not accept it and instead report its current valid value.